requirements.txt
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
torchvision==0.2.2
python-telegram-bot
pillow==8.4.0
numpy==1.19.5
pysocks==1.7.1